tests

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: masnus <masnus@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/21 19:54:40 by masnus            #+#    #+#             */
-/*   Updated: 2024/10/24 21:57:14 by masnus           ###   ########.fr       */
+/*   Updated: 2024/11/09 17:27:41 by masnus           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
